+++ engine/base/src/main/java/provider/wz/WZTool.java
@@ -98,7 +98,7 @@ public class WZTool {
          if (b == 0x7F) {
             strLength = llea.readInt();
          } else {
-            strLength = (int) b;
+            strLength = b;
          }
          if (strLength < 0) {
             return "";
@@ -167,7 +167,7 @@ public class WZTool {
       if (b == -128) {
          return lea.readInt();
       } else {
-         return ((int) b);
+         return b;
       }
    }
